requirements.txt
numpy
pandas
matplotlib
scikit-learn
statsmodels
pmdarima
joblib
tqdm
python-dotenv
jupyter
ipykernel
